# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for a restaurant management system.",
        "title": "Restaurant Management API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/foods": {
            "post": {
                "description": "Create a new food item",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "foods"
                ],
                "summary": "Create food",
                "parameters": [
                    {
                        "description": "Food data",
                        "name": "food",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Food"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.InsertOneResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/foods/{food_id}": {
            "get": {
                "description": "Get a food item by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "foods"
                ],
                "summary": "Get food",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Food ID",
                        "name": "food_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Food"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a food item",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "foods"
                ],
                "summary": "Update food",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Food ID",
                        "name": "food_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Food data",
                        "name": "food",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Food"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.UpdateResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/invoices": {
            "get": {
                "description": "Get all invoices",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invoices"
                ],
                "summary": "List invoices",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.CustomBsonM"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new invoice",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invoices"
                ],
                "summary": "Create invoice",
                "parameters": [
                    {
                        "description": "Invoice data",
                        "name": "invoice",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Invoice"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.InsertOneResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/invoices/{invoice_id}": {
            "get": {
                "description": "Get an invoice by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invoices"
                ],
                "summary": "Get invoice",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invoice ID",
                        "name": "invoice_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.InvoiceViewFormat"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an invoice",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invoices"
                ],
                "summary": "Update invoice",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invoice ID",
                        "name": "invoice_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Invoice data",
                        "name": "invoice",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Invoice"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.UpdateResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.InvoiceViewFormat": {
            "type": "object",
            "properties": {
                "invoice_id": {
                    "type": "string"
                },
                "order_details": {},
                "order_id": {
                    "type": "string"
                },
                "payment_due": {},
                "payment_due_date": {
                    "type": "string"
                },
                "payment_method": {
                    "type": "string"
                },
                "payment_status": {
                    "type": "string"
                },
                "table_number": {}
            }
        },
...
        "model.CustomBsonM": {
            "type": "object",
            "properties": {
                "field1": {
                    "type": "string"
                },
                "field2": {
                    "type": "string"
                }
            }
        },
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "model.Food": {
            "description": "Food item details swagger:model",
            "type": "object",
            "required": [
                "food_image",
                "menu_id",
                "name",
                "price"
            ],
            "properties": {
                "created_at": {
                    "description": "Created timestamp of the food item",
                    "type": "string"
                },
                "food_id": {
                    "description": "Unique ID of the food item",
                    "type": "string"
                },
                "food_image": {
                    "description": "Image URL of the food item\n@swag.Required",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the food item\n@swag.Type string",
                    "type": "string"
                },
                "menu_id": {
                    "description": "ID of the menu associated with the food item\n@swag.Required",
                    "type": "string"
                },
                "name": {
                    "description": "Name of the food item\n@swag.Required",
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "price": {
                    "description": "Price of the food item\n@swag.Required",
                    "type": "number"
                },
                "updated_at": {
                    "description": "Updated timestamp of the food item",
                    "type": "string"
                }
            }
        },
        "model.InsertOneResult": {
            "type": "object",
            "properties": {
                "insertedId": {
                    "type": "string"
                }
            }
        },
        "model.Invoice": {
            "description": "Invoice details swagger:model",
            "type": "object",
            "required": [
                "payment_status"
            ],
            "properties": {
                "Payment_due_date": {
                    "description": "Due date for the payment",
                    "type": "string"
                },
                "created_at": {
                    "description": "Created timestamp of the invoice",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the invoice\n@swag.Type string",
                    "type": "string"
                },
                "invoice_id": {
                    "description": "Unique ID of the invoice",
                    "type": "string"
                },
                "order_id": {
                    "description": "ID of the associated order",
                    "type": "string"
                },
                "payment_method": {
                    "description": "Payment method used for the order\n@swag.Required",
                    "type": "string"
                },
                "payment_status": {
                    "description": "Payment status of the order\n@swag.Required",
                    "type": "string"
                },
                "updated_at": {
                    "description": "Updated timestamp of the invoice",
                    "type": "string"
                }
            }
        },
...
        "model.UpdateResult": {
            "type": "object",
            "properties": {
                "matchedCount": {
                    "type": "integer"
                },
                "modifiedCount": {
                    "type": "integer"
                },
                "upsertedCount": {
                    "type": "integer"
                },
                "upsertedId": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}